Reorganize mock data.
