+++ tests/test.js
@@ -19,7 +19,7 @@ const assert = require('chai').assert;
 const common = require('./test-common');
 const constants = require('../lib/constants');
 const jsigs = require('..');
-const mock = require('./mock-data');
+const mock = require('./mock/mock');
 const {suites} = require('../lib/suites');
 const util = require('../lib/util');
